Tweaks to submodule handling
Add a `git submodule sync`. This allows for the user to change the
origin URL of the submodule and propogate that to change to all hosts.

Change submodule update to not use submodule foreach, which shouldn't be
needed (and also didn't work in my set up).

And an --init to the submodule update call to catch any newly added
submodules.  This is harmless for already init'd submodules.